+++ Examples/AdditionalExamples/AppDelegate.swift
@@ -7,6 +7,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
+        // Request authorization to user notifications.
+        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound, .badge]) { _, _ in
+            // No-op
+        }
         // Override point for customization after application launch.
         return true
     }